restructured controllers
- split between controllers/hardware and controllers/services
