+++ app/config/openscan.py
@@ -1,7 +1,6 @@
 import json
 import pathlib
 import os
-import time
 
 from linuxpy.video.device import iter_video_capture_devices
 import gphoto2 as gp
@@ -9,14 +8,14 @@ from picamera2 import Picamera2
 
 
 from app.config.camera import CameraSettings
-from app.models.camera import Camera, CameraType, CameraMode
+from app.models.camera import Camera, CameraType
 from app.config.cloud import CloudSettings
 from app.config.motor import MotorConfig
 from app.models.motor import Motor, MotorType
 from app.config.light import LightConfig
 from app.models.light import Light, LightType
 
-from app.controllers.cameras.camera import CameraControllerFactory
+from controllers.hardware.cameras import CameraControllerFactory
 
 from dotenv import load_dotenv
