+++ app/controllers/hardware/cameras/linuxpy.py
@@ -1,9 +1,8 @@
-import io
 from tempfile import TemporaryFile
 from typing import IO
 from linuxpy.video.device import Device, VideoCapture
 
-from app.controllers.cameras.camera import CameraController
+from controllers.hardware.cameras.camera import CameraController
 from app.models.camera import Camera, CameraMode
