+++ app/main.py
@@ -2,7 +2,7 @@ import uvicorn
 from fastapi import FastAPI, Body
 from fastapi.middleware.cors import CORSMiddleware
 
-from routers import cameras, motors, projects, cloud, io, paths, scans, focus, lights
+from routers import cameras, motors, projects, cloud, io, paths, scans, lights
 
 app = FastAPI()
 
@@ -20,7 +20,6 @@ app.include_router(lights.router)
 app.include_router(projects.router)
 app.include_router(io.router)
 app.include_router(scans.router)
-app.include_router(focus.router)
 
 app.include_router(cloud.router)
 app.include_router(paths.router)
